fix: parseFloat para precios, pre-check sesión, featured en inventory_cakes
- server.js: parseInt → parseFloat en variants.price al agregar producto
- admin-stock-pasteles.js: verifica /api/auth/me antes de POST para dar
  feedback claro ("Sesión expirada") en vez de un 401 silencioso
- inventory_cakes.json: featured:true en productos 8 y 9

+++ server/server.js
@@ -1930,7 +1930,7 @@ app.post('/api/inventory/cakes/add', requireAdmin, (req, res) => {
     sizes: {},
     variants: variants.map(v => ({
       name: String(v.name || '').trim(),
-      price: Math.max(0, parseInt(v.price) || 0),
+      price: Math.max(0, parseFloat(v.price) || 0),
       enabled: v.enabled !== false
     })).filter(v => v.name)
   };
